Move field finalization to ^FS
